20240323_윤덕중 payment 전체 완료. 중간 push

+++ myEduMySelect/src/main/java/com/myedumyselect/admin/openapi/data/controller/DataController.java
@@ -1,17 +1,15 @@
 package com.myedumyselect.admin.openapi.data.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.ApplicationArguments;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.SessionAttribute;
 
 import com.myedumyselect.admin.login.vo.AdminLoginVO;
 import com.myedumyselect.admin.openapi.data.service.DataService;
 
-import jakarta.servlet.http.HttpSession;
 import lombok.Setter;
 
 @RestController
@@ -22,11 +20,11 @@ public class DataController {
 	private DataService dataService;
 
 	@GetMapping(value = "/insertAcademySource", produces = "application/json; charset=UTF-8")
-	public String insertAcademySource(HttpSession session) throws Exception {
-//		AdminLoginVO adminLoginVO = (AdminLoginVO) session.getAttribute("adminLogin");
-//		if (adminLoginVO == null) {
-//			return "redirect:/admin/login";
-//		}
+	public String insertAcademySource(
+			@SessionAttribute(name = "adminLogin", required = false) AdminLoginVO adminLoginVO) throws Exception {
+		if (adminLoginVO == null) {
+			return "redirect:/admin/login";
+		}
 		dataService.deleteAcademySourceList();
 		int listTotalCount = dataService.listTotalCount();
 		dataService.insertAcademySourceList(listTotalCount);
@@ -35,7 +33,11 @@ public class DataController {
 	}
 
 	@GetMapping(value = "/getAcademySourceCount", produces = MediaType.APPLICATION_JSON_VALUE)
-	public String getAcademySourceCount() throws Exception {
+	public String getAcademySourceCount(
+			@SessionAttribute(name = "adminLogin", required = false) AdminLoginVO adminLoginVO) throws Exception {
+		if (adminLoginVO == null) {
+			return "redirect:/admin/login";
+		}
 		int totalCount = dataService.listTotalCount();
 		int currCount = dataService.listCurrCount();
